fix(services): Eliminate 114 analyzer violations in BotCore Services - Phase 2 surgical fixes (#494)
* Initial plan

* fix(services): CA1307, CA1024, CA1836, S1144 - string comparisons, property conversions, collection checks, unused methods


* fix(services): CA1034, S108, CA1033 - nested types, empty blocks, sealed class


* fix(services): Eliminate 114 analyzer violations in BotCore Services - Phase 2 surgical fixes


---------

+++ src/BotCore/Services/AutonomousDecisionEngine.cs
@@ -1603,7 +1603,7 @@ public class AutonomousDecisionEngine : BackgroundService
     private static double CalculateVolumeMA(List<Bar> bars, int period)
     {
         if (bars.Count < period) return 0;
-        return (double)bars.TakeLast(period).Select(b => b.Volume).Average();
+        return bars.TakeLast(period).Select(b => b.Volume).Average();
     }
     
     private static decimal CalculateEMA(List<decimal> values, int period)
